+++ CHANGELOG.md
@@ -5,6 +5,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Private function:
+  - `Invoke-ReportServerSetupAction`
+
 ### Changed
 
 - SqlServerDsc
@@ -26,6 +31,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Fix localization strings in `Assert` method.
 - `Save-SqlDscSqlServerMediaFile`
   - Fix localizations strings that used wrong keys.
+  - Fix unit tests so the work cross-platform.
+- `Install-SqlDscServer` and private function `Invoke-SetupAction`
+  - Fix localization string keys naming.
+  - Fix unit tests to use correct localization string names.
 - `SqlConfiguration`
   - Change the alias command to real command name, to pass HQRM tests.
 - `SqlDatabaseUser`